Make folder with from-ps stuff
I set it up so it takes the md file and if you use the generate_from_md.sh on linux
(open it in a text-editor to see which packages need to be installed).

It still needs to be updated, but being able to generate this is useful.

CCMAIL:kimageshop@kde.org