Cleanup: Get all of the old stuff out of the way, for now.
I'm trying to wrap my brain around the structure of what I'm doing,
and I can't easily do that with a lot of random files in the way.

There are probably some useful things here still, but for now I'm just
going to move everything out of the way so I can focus on the task at hand.